Update chill psycopg2 dependency
Should use the libpq-dev package to install the pg_config binary that
the python psycopg2 requires in 2.8.0.  Pin this dependency down to <2.9
to avoid future breaking changes.

Note that chill and puzzle-massive isn't using postgresql yet. It is a planned
feature to transition off of sqlite if the performance is better.