Fix janitor, include a cleanup flag
The convertPiecesToDB process that the janitor would do will now work
outside of the application context.  Redis data will now properly
transfer to the DB when needed.  This may fix the pieces group bug #63
since previous redis data from a completed puzzle could have been reused
when the puzzle was rebuilt.

Added a cleanup flag for the backup-db.sh script.  This option should be
used after stopping the app.  The backup-db.sh can be used while the
site is running, but is not recommended.  The cleanup option shouldn't
be used when the site is running.