Improve transfer of redis piece data to db
Only execute the SQL update for Piece data if the redis piece data has
changed when running convertPiecesToDB.transfer. The update is also
skipped if no piece data is in redis.
The change to piece parent/g and status/s shouldn't cause any piece
movement bugs. The redis piece data is always removed when a puzzle is
rebuilt.