Improve the purge process of cache url list
Avoids race conditions when the file is modified while it is being
processed.  Should work better since the file is first copied to a tmp
directory and then removed before it takes the time to process. Waiting
a second and then re-running the process helps to catch any new changes
that were missed. It isn't perfect, but seems to work for this use case.