Remove unneeded files
Converted the Procedures doc and Test-Plan to docs/todo/ directory in
a more parseable format.  The content for these can then be later
integrated in with the project as entries in the change log and existing
documentation.

The out/ directory was removed since it looked like it was just
generated from jsdoc and had no content.

The added files to the bin/ directory were also removed since they are
generated.